move serde helpers to separate file
