+++ src/lib.rs
@@ -49,6 +49,7 @@ mod gtfs;
 mod gtfs_reader;
 pub(crate) mod objects;
 mod raw_gtfs;
+mod serde_helpers;
 
 #[cfg(test)]
 mod tests;
